requirements.txt
openai>=2.26.0        # OpenAI Python SDK (interpreter uses GPT-4.1)
pydantic>=2.0.0          # schema validation for interpreter output and domain models
python-dotenv>=1.0.0     # loading OPENAI_API_KEY from .env
pytest>=8.0.0            # test runner
pytest-asyncio>=0.23.0   # async test support
rich>=13.0.0             # pretty CLI output for benchmark runner